rte_2071759301 rename file when old file wasn't set caused NullPointerException
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@20098 0c594bb0-8543-0410-a25e-d8ec440261db

+++ VirboAutoplot/src/org/virbo/autoplot/scriptconsole/ScriptPanelSupport.java
@@ -277,11 +277,13 @@ public class ScriptPanelSupport {
                     file = new File(file.toString() + ".jy");
                 }
             }
-            if ( cb.isSelected() && !oldFile.equals(file) ) {
-                if ( ! oldFile.delete() ) {
-                    JOptionPane.showMessageDialog( panel, "unable to delete old file: "+oldFile );
+            if ( oldFile!=null ) {
+                if ( cb.isSelected() && !oldFile.equals(file) ) {
+                    if ( ! oldFile.delete() ) {
+                        JOptionPane.showMessageDialog( panel, "unable to delete old file: "+oldFile );
+                    }
                 }
-            }            
+            }
         }
         return r;
     }
